bigint_t.size -> size_t
git-svn-id: https://srv183.svn-repos.de/dev34/pdclib/trunk@989 bcf39385-58cc-4174-9fcf-14f50f90dd47

+++ functions/_PDCLIB/_PDCLIB_bigint2.c
@@ -8,9 +8,11 @@
 
 #include "pdclib/_PDCLIB_internal.h"
 
+#include <stddef.h>
+
 _PDCLIB_bigint_t * _PDCLIB_bigint2( _PDCLIB_bigint_t * bigint, unsigned n )
 {
-    int i;
+    size_t i;
 
     /* Check which bigint digit will contain the set bit */
     bigint->size = ( n / _PDCLIB_BIGINT_DIGIT_BITS ) + 1;
